Convert omerofeatures.py into a proper script
